Checkpoint - 2021-02-02
Working on the 'new' superclass for a configparser wrapper.

A couple things here, splitting the `key` into `op1` and `op2` pairs
is trickier than in the original `setenvironment` class because I wanted
to be able to support op2 parameters that might have spaces.

For example, all of these 'key' entries should resolve to
`op1 = AAA` and `op2 = BBB`:
- `AAA`  (op2 should be None)
- `AAA BBB`
- `AAA 'BBB'`
- `AAA 'BB B'` (op2 should be `BB B` in this case, quote not included in
  the resulting string).
- `AAA BBB CCC`
- `AAA 'BBB' CCC`
- `AA-A BBB` (op1 should be changed to `AA_A` because of how the handler
  is working.

The new parser handler system in this class works like this:
- when processing `op1` the parser will look for a class method named
  `_handler_<op1>(section_name, op1, op2, data,
      processed_sections=None)`. If it finds one then it will call that
  method and pass in the appropriate parameters.
- If it does not find a handler for `op1` then it just passes that
  key/value pair up and doesn't do anything with it.
- Because we must look for a valid generated method name in our class,
  we need to replace any 'bad' characters in the `op1` field. In this
  case, our regex will attempt to replace a dash (`-`) with an
  underscore (`_`).

TODO: This is still a WIP.
- We might consider replacing the regex
expression from text with a property so derivitive classes could easily
override it by changing the property.
  - related to this, might create some kind of a `normalize_op1()`
    method, that could be also overridden or something to deal with
    operand clanup operations?