Initial ThrowControl class
Implemented a 'baseclass' called `ThrowControl`
which probably needs to be renamed.

The idea of this class is that we inherit from it and can then
set up conditional error messages that will either throw an error
based on the severity of the error and threshold parameter or if
it will just print out a warning message + traceback or be silent.