Refactor phase 1
Changing a lot of names of properties to be 'private' members
because I want these things to be thought about before someone
just overrides them.  This also makes the public interface more
explicit.

Next steps are to modify the parser to accept handlers using the
naming scheme `_handler_<foo>` or `handler_<foo>` which should
be useful to designate what _can_ and what _cannot_ be overridden.