Partial Overhaul - Cleaning up for Production
Updates to documentation

Renamed a number of members of various classes to make it
more explicit what handlers, etc. are _allowed_ to be overridden
etc.

Added new capability to load in a handler that conforms to
`_handler_<operation>` or `handler_<operation>` to better
support the differentiation between what is seen as 'internal'
(and thus really shouldn't be overridden) and what is seen as
'user api' and thus may be overridden.

Todo:
- Testing needs to be updated.
- SetEnvironment needs to be completed.
- Add some mechanism to alert if there are both public and private
  handlers defined for the same operation (public one wins?)