ENHANCE: Match configparser behaviour for No Value
`configparser` can be set to allow section options that don't have a
colon or equals if there is no value... but the behaviour is different
in terms of what the results are if there is or isn't that separator
provided.

For example:

```ini
[SEC]
key1:
key2
key3:
```

will result in key:value pairs for key1 and key3 to have an empty string
as the value. `key2` however will have `None` for its value field.

This commit brings the behaviour of ConfigParserEnhanced into line with
what ConfigParser does in this case.