Renamed `handler_generic` to avoid accidental use
The 'generic' handler should not be named "handler_generic" because this
could cause the parser to use it if someone had a .ini file that looks
like this:

```ini
[SECTION]
generic commands are cool: but really, this shouldn't invoke a handler.
```

It still serves as a fantastic handler template but the naming shouldn't
conflict with the pattern since it's really kind of a non-handler.

Alternatively, we could create a 'handler function map' that would map
operations to handlers... buuuut I don't love that because subclasses
would then need to have some awareness of superclass handlers, etc.