ENHANCE: Add normalization helpers for `operation` and `parameter`
Added a `_apply_transformation_to_operation` and
`_apply_transformation_to_parameter` helpers which can be used to apply
transformations to the fields as needed.  For now the only thing being
done is to replace `-` with `_` in _operation_ entries within the
handler parameters. This should simplify life a bit further downstream
since we use map these operations over to things like function names
etc.