Fix issue with slashes in envvar values during 'envvar-remove'
The problem is that if the character '#' appears in the string
it might cause issues now because we changed the character SED
is using as the separator.

Is there some other character that might be useful? We might need to
do a replacement of the 'bad char' first to some kind of string, say
something like '=|=' (that would have a low likelihood of happening
... or maybe ':::'). We might also need to look into something other
than SED down in the internals as well.