Remove comment
Note that the SERIAL_* environment variables look like they're only
defined on RHEL7; that is, the `no-mpi` configuration looks like it's
only supported on RHEL7.  What we have here will match the existing
implementation, so I don't know that there's more to do in terms of
checking to make sure the user knows what they're doing.