ISSUE 13: Dealing with DEFAULT sections
This patch adds a `_default_section_name` property which controls the
underlying `ConfigParser` representation of the "DEFAULT" section.
We default it to "CONFIGPARSERENHANCED_COMMON" for now.
This property may need to be renamed to something more appropriate or
we should remove the property alltogether and just hard-code that into
ConfigParserEnhanced and document it as something that one should avoid
using? Alternatively we could add checks to flag an error _if_ we detect
that a .ini file actually contains that section so it's not used since
it can cause so much undefined behavior.