This adds the -fPIC  flag at the start of the CXX flags
Why do this? If we build as shared and -fPIC is after -O3 the link fails.
If we use the CMAKE_POSITION_INDEPENDENT_CODE boolean the -fPIC flag
gets set at the end. I belive this is a flaw in the compiler/linker
set so this should just avoid it.